# Parallel parsing