--- a/Cargo.toml
+++ b/Cargo.toml
@@ -25,6 +25,7 @@ prost-types = "0.3"
 stderrlog = "0.4"
 structopt = "0.2"
 rayon = "1.0.3"
+futures = "0.1"
 
 [build-dependencies]
 prost-build = "0.3"