--- a/src/main.rs
+++ b/src/main.rs
@@ -20,11 +20,13 @@ extern crate rayon;
 extern crate stderrlog;
 #[macro_use]
 extern crate structopt;
+extern crate futures;
 
 mod args;
 mod ids;
 mod osmflat;
 mod osmpbf;
+mod parallel;
 mod stats;
 mod strings;
 
@@ -43,7 +45,6 @@ use std::collections::{hash_map, HashMap};
 use std::fs::File;
 use std::io::{self, Read, Seek};
 use std::str;
-use std::sync::{mpsc::sync_channel, Arc, Condvar, Mutex};
 
 fn serialize_header(
     header_block: &osmpbf::HeaderBlock,
@@ -445,56 +446,29 @@ fn run() -> Result<(), Error> {
     // Serialize dense nodes
     let mut nodes_id_to_idx = ids::IdTableBuilder::new();
     if let Some(index) = pbf_dense_nodes {
-        let index: Vec<_> = index.collect();
-        let index_len = index.len();
-
-        let mut pb = ProgressBar::new(index_len as u64);
-        pb.message("Converting dense nodes...");
         let mut nodes = builder.start_nodes()?;
-        let iter = Arc::new(Mutex::new(index.into_iter().enumerate()));
-        let next = Arc::new((Mutex::new(0_usize), Condvar::new()));
-        let (sender, receiver) = sync_channel(rayon::current_num_threads());
-        for _ in 1..rayon::current_num_threads() {
-            let mut local_file = File::open(&args.input)?;
-            let iter = iter.clone();
-            let next = next.clone();
-            let sender = sender.clone();
-            rayon::spawn(move || loop {
-                let idx = iter.lock().unwrap().next();
-                let idx = match idx {
-                    Some(x) => x,
-                    None => break,
-                };
-
-                let block: io::Result<osmpbf::PrimitiveBlock> = read_block(&mut local_file, &idx.1);
-
-                let mut guard = next.0.lock().unwrap();
-                while *guard != idx.0 {
-                    guard = next.1.wait(guard).unwrap();
-                }
-                sender.send((idx.0, block)).unwrap();
-                *guard += 1;
-                next.1.notify_all();
-            });
-        }
-        for idx in 0..index_len {
-            let block = receiver.recv().unwrap();
-            assert_eq!(idx, block.0);
-            stats += serialize_dense_nodes(
-                &block.1?,
-                &mut nodes,
-                &mut nodes_id_to_idx,
-                &mut stringtable,
-                &mut tags,
-            )?;
-            pb.inc();
-        }
+        let index: Vec<_> = index.collect();
+        parallel::parallel_process(
+            "Converting dense nodes...",
+            index.into_iter(),
+            || -> Result<File, Error> { Ok(File::open(&args.input)?) },
+            |mut file, idx| read_block(&mut file, &idx),
+            |data| -> Result<(), Error> {
+                stats += serialize_dense_nodes(
+                    &data?,
+                    &mut nodes,
+                    &mut nodes_id_to_idx,
+                    &mut stringtable,
+                    &mut tags,
+                )?;
+                Ok(())
+            },
+        )?;
         {
             let mut sentinel = nodes.grow()?;
             sentinel.set_tag_first_idx(tags.next_index());
         }
         nodes.close()?;
-        pb.finish();
     }
     info!("Dense nodes converted.");
     let nodes_id_to_idx = nodes_id_to_idx.build();