On my machine 240% faster planet parsing

Now:
1467.69s real, 3252.31s user, 503.77s sys, 58374192Kb Max RSS

Before:
 3296.77s real, 2993.31s user, 159.79s sys, 57005004Kb Max RSS